+++ src/BAT.jl
@@ -44,6 +44,7 @@ import Measurements
 import NLSolversBase
 import Optim
 import AdvancedHMC
+import Bijectors
 
 #for AHMI
 using DataStructures
