+++ src/samplers/mcmc/ahmc/ahmc.jl
@@ -5,4 +5,5 @@ include("ahmc_integrator.jl")
 include("ahmc_proposal.jl")
 include("ahmc_metric.jl")
 include("ahmc_sampler.jl")
+include("only_burnin_tuner.jl")
 #TODO@C #include("ahmc_iterator.jl")
